requirements.txt
Flask==2.0.1
gunicorn==20.1.0
numpy==1.19.5
Pillow==8.4.0
Jinja2==3.0.1
urllib3==1.20.0
Werkzeug==2.0.1
ultralytics==8.0.20
opencv-python-headless==4.5.5.64
protobuf==3.20.*
